Rename `abilities.ts` to `all-abilities.ts`
