+++ src/data/ab-attrs/weather-based-speed-doubler-ab-attr.ts
@@ -4,7 +4,7 @@ import type { BattleStat } from "#enums/stat";
 import { Stat } from "#enums/stat";
 import type { Pokemon } from "#app/field/pokemon";
 import type { NumberHolder } from "#app/utils";
-import { getWeatherCondition } from "../abilities";
+import { getWeatherCondition } from "#app/data/all-abilities";
 
 /**
  * Ability attribute that doubles speed if specific weather(s) are active
